Avoid SubfieldBase deprecation warning in Django >= 1.8
Starting with Django 1.8, JSONField will no longer JSON-decode
strings that you assign to the field. Instead the strings will be
considered as the value to store and they will be JSON-encoded
in the database.

To avoid breaking applications running in the live, JSONField
keeps doing that automatic JSON decode with older Django versions.