Improve jsonify to limit the risk of XSS attacks
This is a backwards incompatible change as users of jsonify in <script>
tags will have to add the safe filter. Since we released 1.0.0 recently,
we could decide it's part of the big interface change that justified the
increase of the major number.

Inspired by various code snippets found in
https://code.djangoproject.com/ticket/17419

Fix #47